Rename file names using the already given convention
